all tests invoking external binaries considered integration tests to satisfy wercker
